% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confset.R
\name{ConfSet}
\alias{ConfSet}
\title{Konfidenz-Obermenge eines Excursion Sets}
\usage{
ConfSet(data, S, partitions, alpha, level, pmethod = "tgkf", mb.iter = 1000)
}
\arguments{
\item{data}{Realisierungen des Zufallsfeldes.}

\item{S}{Grundmenge S (1D und EC(S) = 1).}

\item{partitions}{Partitionierung der Grundmenge (degenerierend und verfeinernd).}

\item{alpha}{Signifikanzniveau für die Konfidenzmenge.}

\item{level}{Grenzwert des Excursion Sets.}

\item{pmethod}{Approximationsmethode der p-Werte:
"tgkf" - Gaußsche Kinematische Formel für t-Verteilungen (default),
"mboot"- Multiplier Bootstrap.}

\item{mb.iter}{Anzahl der Iterationen für den Multiplier Bootstrap (default: 1000).}
}
\value{
Konfidenzmenge U mit \eqn{P(S_0 \subset U) \geq 1 - \alpha}
}
\description{
Berechnung einer Konfidenz-Obermenge des Excursion Sets
\eqn{{s: \mu(s) > level}} bzgl. eines Zufallsfelds
\eqn{X(s) = \mu(s) + \sigma(s)\epsilon(s)} zu einer kompakten Grundmenge S mit
Eulercharakteristik 1.
Die Berechnung basiert auf der Metode aus
"False Discovery Control for Random Fields" (2004) - M. Perone Pacifico,
C. Genovese, I. Verdinelli and L. Wasserman
}
