% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{partition_seq}
\alias{partition_seq}
\title{Degenerative Partitionsfolge}
\usage{
partition_seq(S, pieces = 2, max_iterations = 1000)
}
\arguments{
\item{S}{Grundmenge, die partitioniert werden soll.}

\item{pieces}{Anzahl der Teile, in die gesplittet werden soll.}

\item{max_iterations}{maximale Anzahl an Folgengliedern.}
}
\value{
Eine Liste je Folgenglied. Jede dieser Listen enthält die
immer feiner werdende Aufspaltungen der Menge S.
}
\description{
Erzeugt eine degenerative Folge von Partitionen, die immer feiner werden,
in dem die Mengen stets in eine fixe Anzahl an gleichgroßen Teilmengen
geteilt wird.
}
\examples{
partition_seq( seq(0, 1, length.out = 100), pieces = 3 )

}
