% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_supstatistic.R
\name{tGKF}
\alias{tGKF}
\title{Gaussian Kinematic Formular für eindimensionale t-verteilte Zufallsfelder}
\usage{
tGKF(Set, X, threshold)
}
\arguments{
\item{Set}{Grundmenge des Zufallsfeldes.}

\item{X}{Realisierungen des Zufallsfeldes, wobei in den Zeilen die Werte
pro Punkt der Grundmenge und in den Spalten die Werte pro Realisierung
enthalten sind.}

\item{threshold}{Level des Exkursion Sets.}
}
\value{
Die erwartetete Eulercharakteristik des Excursion Sets \eqn{S_0}.
}
\description{
Berechnung der erwarteteten Eulercharakteristik per GKF für das Excursion
Set \eqn{S_0 = {\mu(s) > level}} eines eindimensionalen Zufallsfeldes
zu einem fixen Level.
Die Grundmenge des Zufallsfeldes muss zusammenhängend sein,
also eine Eulercharakteristik von 1 besitzen
}
